fix: update/saving functionality of the base `Settings` class
* Fix key checks for properties
This solves the problem related to how the properties where returned. Instead of checking for a value which represent a property/setting in the class, we were using the old way, back when we cast the value.
* Refactor `DynamicSettings` class to use the `Settings` fixed approach to handle properties
The commit also includes a "regenerate" method, the only intent of this is to create a new instance which will trigger the dynamic filling of properties. This is helpful because can be used to test the `Schema` functionality to "regenerate" the class with the newly created settings.
* Update fake method to return `DynamicSettings` instance
* Add test for `Settings` class
This will use the `DynamicSettings` class which only purpose is to generate dynamic properties in a `SettingsClass` for testing, but the hydration and use of the `Settings` base class is the same.
* Refactor `SchemaTest` to use the newly updated approach to test settings dynamically
* Refactor code to match return types and fix code style issues